Image_Text_Gemini_Bot/requirement.txt
google-generativeai
streamlit
Pillow
requests